Streamline handling of recovery vs regular blocks
Incorporate information about whether the block is a regular or
recovery block into the block metadata type, and keep track of this
information in a "block catalog." This is currently maintained on the
fileserver but it could be moved to individual peers to allow them to
keep track of their own files.

This also unifies RedundantStorageScheme and ErrorRecoveryScheme by
considering that any scheme that adds redundancy to the data with the
goal of recovering from errors and erasures is an error recovery
scheme.